Detect left and right line
Previously, there can be several lines overlapping with each other. These lines makes the result image messy. To make the lane detection clean, I applied average to the lines and detect a set of left lines and a set of right lines.
By averaging the m and b parameter of each line, I comes up with one line represents the left line and one represents the right line.
I extended the line all the way to the bottom to make it look good.

Also fixed a few bugs and added some test results